day_11_agentic_rag/requirements.txt
# Core LangChain (updated for compatibility)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0

# Vector Stores
faiss-cpu>=1.7.4
chromadb>=0.4.22
lancedb>=0.4.0

# Document Processing
pypdf>=3.17.4
python-docx>=1.1.0
unstructured>=0.11.8

# Embeddings
sentence-transformers>=2.2.2
langchain-huggingface>=1.0.0
tiktoken>=0.5.2

# LLM Providers
openai>=1.6.1
anthropic>=0.8.1

# Web UI
streamlit>=1.29.0
gradio>=4.13.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
numpy>=1.26.2
pandas>=2.1.4

